GUI and beginning of optimization
I have now created a basic GUI for playing the game Sudoku, where the player gets a blank board and first fills in the fixed values of their sudoku, before they use the fix button and can begin actually solving the sudoku. I also started working on the part where the machine solves the sudoku for the player if that is wanted. This part will be based on the following article: https://www.mn.uio.no/math/english/research/projects/focustat/the-focustat-blog%21/sudokustory.html